gitops/pyproject.toml
[project]
name = "gitops"
version = "0.1.0"
description = "Gitops using ADK"
authors = [{ name = "Karankumar Adep", email = "karankumar.adep@dexcom.com" }]
license = "Apache-2.0"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "google-adk==1.3.0",
    "python-dotenv==1.1.0",
    "requests==2.31.0",
    "google-cloud-aiplatform[agent_engines]>=1.91.0,!=1.92.0",
    "absl-py>=2.2.1,<3.0.0",
    "pydantic>=2.10.6,<3.0.0",
]
